TransactionInfo re-impl.
