+++ modules/explorer-api/src/main/scala/org/ergoplatform/explorer/http/api/v0/defs/AddressesEndpointDefs.scala
@@ -1,10 +1,10 @@
 package org.ergoplatform.explorer.http.api.v0.defs
 
 import org.ergoplatform.explorer.http.api.ApiErr
-import org.ergoplatform.explorer.{Address, TokenId}
-import org.ergoplatform.explorer.http.api.models.{Items, Paging}
 import org.ergoplatform.explorer.http.api.commonDirectives._
+import org.ergoplatform.explorer.http.api.models.{Items, Paging}
 import org.ergoplatform.explorer.http.api.v0.models.{AddressInfo, TransactionInfo}
+import org.ergoplatform.explorer.{Address, TokenId}
 import sttp.tapir._
 import sttp.tapir.json.circe._
